// src/app/shared/model/produit-simple/inst-caract-var.model.ts
export class InstCaractVar {
  _id?: string;
  refInstCaractVar?: string;
  affUniqPriv?: boolean;
  groupe?: string;
  label?: string;
  valeur?: string;
  unite?: string;
  ordre?: number;
  coutAdd?:boolean;
  typeCoutAdd?: string;
  valeurCoutAdd?: number;
  prix?: number;
  produitService?: any;
  indPromo?: any[];
  indFraisAdd?: any[];
  indStock?: any[];
  expedition?: any[];
}